Updated Namespace to make it a CompareAsiaGroup library and added changes to README.md
