+++ README.md
@@ -8,14 +8,18 @@ Registry manager for Laravel 5. An alternative for managing application configur
 
 ## Installation
 
-- [Registry on Packagist](https://packagist.org/packages/torann/registry)
-- [Registry on GitHub](https://github.com/Torann/laravel-registry)
-- [Laravel 4 Installation](http://lyften.com/projects/laravel-registry/doc/laravel-4.html)
+- [Registry on GitHub](https://github.com/compareasia/laravel-registry)
 
 To get the latest version of Registry simply require it in your `composer.json` file.
 
 ~~~
-"torann/registry": "0.2.*@dev"
+"repositories": [
+	  {
+	  	"type": "vcs",
+	  	"url": "https://github.com/compareasiagroup/laravel-registry"
+		}
+  	],
+"require" : "compareasiagroup/laravel-registry": "dev-master"
 ~~~
 
 You'll then need to run `composer install` to download it and have the autoloader updated.
@@ -24,7 +28,7 @@ Once Registry is installed you need to register the service provider with the ap
 
 ```php
 'providers' => array(
-    'Torann\Registry\RegistryServiceProvider',
+    'CompareAsiaGroup\Registry\RegistryServiceProvider',
 )
 ```
 
@@ -32,7 +36,7 @@ Registry also ships with a facade which provides the static syntax for creating 
 
 ```php
 'aliases' => array(
-    'Registry' => 'Torann\Registry\Facades\Registry',
+    'Registry' => 'CompareAsiaGroup\Registry\Facades\Registry',
 )
 ```
 
@@ -44,11 +48,11 @@ Run this on the command line from the root of your project:
 $ php artisan vendor:publish
 ~~~
 
-A configuration file will be publish to `config/registry.php` and a migration file to `database/migrations/`
+A configuration file will be publish to `config/registry.php`
 
 ## Documentation
 
-[View the official documentation](http://lyften.com/projects/laravel-registry/).
+@TODO Add documentation to Confluence
 
 ## Change Log
